backend/requirements.txt
# Django Core
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0

# Database
mysqlclient==2.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# AI/ML Libraries
google-generativeai>=0.3.0
PyPDF2==3.0.1

# LiveKit Integration
livekit-agents[bey]~=1.2
livekit-plugins-google>=1.2.0
livekit-plugins-openai>=1.2.0
livekit-plugins-deepgram>=1.2.0

# Utilities
python-dotenv==1.0.0
python-decouple==3.8
requests==2.31.0
aiohttp==3.9.0
Pillow==10.0.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0